move all the itemloader benchmarker to one file
