--- itemloader/itemloader/settings.py
@@ -1,9 +1,0 @@
-BOT_NAME = 'itemloader'
-
-SPIDER_MODULES = ['itemloader.spiders']
-NEWSPIDER_MODULE = 'itemloader.spiders'
-
-ROBOTSTXT_OBEY = True
-
-LOG_LEVEL = 'INFO'
-LOGSTATS_INTERVAL = 1
